+++ packages/ui-android/src/index.ts
@@ -1,3 +1,5 @@
 export { default as BottomNav } from './components/BottomNav';
 export { default as UpdateIndicator } from './components/UpdateIndicator';
 export { default as UpdateDiagnosticsSheet } from './components/UpdateDiagnosticsSheet';
+export { default as StageCorePanel } from './components/StageCorePanel';
+